Ng2 Migration: Keep moving code from AngularOrcidGenerated to its own files
